feat: translations and filtering relationship types
