+++ packages/esm-patient-clinical-view-app/translations/en.json
@@ -5,6 +5,8 @@
   "activeCases": "Active cases",
   "add": "Add",
   "addCase": "Add case",
+  "addPNSContact": "Add PNS Contact",
+  "addRelationship": "Add relationship",
   "admissionDate": "Admission Date",
   "admissionWard": "Admission Ward",
   "admittingDoctor": "Admitting Doctor",
@@ -55,7 +57,7 @@
   "encounters": "",
   "encountersList": "Encounters list",
   "facility": "Facility",
-  "familyHistory": "Family history",
+  "familyContacts": "Family contacts",
   "feedingOrders": "Feeding Orders",
   "fetalHeartRate": "Fetal Heart Rate",
   "finalDiagnosis": "Final Diagnosis",
@@ -105,8 +107,6 @@
   "primaryDiagnosis": "Primary Diagnosis",
   "priorityOfAdmission": "Priority Of Admission",
   "recommendedProcedure": "Recommended Procedure",
-  "recordContact": "Record Contact",
-  "recordHistory": "Record History",
   "recordLabourDetails": "Record labour details",
   "referrals": "Referrals",
   "relation": "Relation",
